Uses execute over remote_file
There is an issue with the remote_file command in Chef 11.4.4 that
doesn't allow us to use local files.

The `create_url` method:

https://github.com/opscode/chef/blob/84e3ac9da/lib/chef/rest.rb#L139-L145

Which is used by the action_create of the chef/provider/remote_file.rb
class:

https://github.com/opscode/chef/blob/84e3ac9da/lib/chef/provider/remote_file.rb#L46

Doesn't properly handle the file:// URI scheme.  Using a straight
execute is the best option here.