Uses not_if to determine if install is needed
execute#creates is run before the not_if block can be executed, and will
and will skip running the command if the file exists.  Since that can't
be helped, I have removed the file and replaced it with the same
functionality in the not_if block.

This also adds a version check to the installed version to make sure if
the versions don't match, we go forth with the install.